More refactorings, of course Adds a weapon_def and a weapon instance. Should these be cloned? Should they reference the same instance? What is the Rc<T> in Rust?
Adds a shooting system.
