--- src/components/control/mod.rs
@@ -1,4 +1,0 @@
-use bevy::prelude::Component;
-
-#[derive(Component, Clone)]
-pub struct TriggerPulled {}
