+++ src/main.rs
@@ -7,15 +7,17 @@ use bevy_rand::plugin::EntropyPlugin;
 use bevy_xpbd_2d::prelude::*;
 use big_brain::{BigBrainPlugin, BigBrainSet};
 use systems::*;
-use crate::components::{Health, PlayerControl, QuadCoord, QuadStore};
+use crate::components::{Health, QuadCoord, QuadStore};
 use rand_chacha::ChaCha8Rng;
 use boids::ai::{attack_and_eat_action_system, find_prey_action_system, Hunger, hunger_scorer_system, hunger_system, hunt_prey_action_system, HuntTarget};
 use boids::components::{BoidDirection, BoidStuff};
 use boids::systems::{boid_steering, quad_boid_flocking, spawn_boids};
+use components::control::PlayerControl;
 use systems::camera::camera_follow;
 use systems::input::{add_mouse_aim_line, draw_mouse_aim, keyboard_input, mouse_look, mouse_position};
 use systems::movement::{linear_velocity_control_boid, linear_velocity_control_player};
-use systems::startup::{load_background, spawn_camera, spawn_player};
+use systems::player::spawn_player;
+use systems::startup::{load_background, spawn_camera};
 
 mod components;
 mod systems;
