+++ src/systems/movement.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::{Query, With, Without};
 use bevy_xpbd_2d::components::{ExternalForce, LinearVelocity};
 use crate::boids::components::{Boid, BoidDirection};
-use crate::components::PlayerControl;
+use crate::components::control::PlayerControl;
 use crate::components::player::Player;
 
 #[allow(dead_code)]
