+++ src/systems/startup.rs
@@ -3,7 +3,7 @@ use bevy::asset::AssetServer;
 use bevy::math::{Rect, Vec2, Vec3};
 use bevy::core::Name;
 use bevy::render::camera::ScalingMode;
-use crate::components::{GameCam, PlayerBundle};
+use crate::components::GameCam;
 use crate::{CAMERA_SCALE, METERS_PER_PIXEL, PIXELS_PER_METER};
 
 pub fn load_background(
@@ -26,29 +26,6 @@ pub fn load_background(
                 texture: asset_server.load("background/background.png"),
                 ..default()
             },
-        ));
-}
-
-pub fn spawn_player(
-    mut commands: Commands,
-    asset_server: Res<AssetServer>) {
-    commands
-        .spawn((
-            SpriteBundle {
-                transform: Transform::from_xyz(
-                    0.0,
-                    0.0,
-                    1.0,
-                )
-                    .with_scale(Vec3::new(
-                        METERS_PER_PIXEL,
-                        METERS_PER_PIXEL,
-                        1.0,
-                    )),
-                texture: asset_server.load("sprites/person.png"),
-                ..default()
-            },
-            PlayerBundle::default(),
         ));
 }
